# frontend/api_doc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Movie Review System API",
    "description": "Backend API for Movie Review Website",
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
    "/user/avatar": {
      "put": {
        "tags": [
          "用户管理"
        ],
        "summary": "更新用户头像",
        "description": "更新当前登录用户的头像 URL",
        "operationId": "updateAvatar",
        "parameters": [
          {
            "name": "avatarUrl",
            "in": "query",
            "description": "头像图片的访问 URL",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/register": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "用户注册",
        "description": "注册新用户，需要提供账号、密码、昵称等信息",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/user/refresh": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "刷新 Token",
        "description": "使用 Refresh Token 换取新的 Access Token",
        "operationId": "refreshToken",
        "parameters": [
          {
            "name": "refreshToken",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/user/logout": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "用户注销",
        "description": "注销当前用户，将 Token 加入黑名单",
        "operationId": "logout",
        "parameters": [
          {
            "name": "refreshToken",
            "in": "query",
            "description": "Refresh Token",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "用户登录",
        "description": "使用账号和密码登录，成功返回 Token 和用户信息",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultUserVO"
                }
              }
            }
          }
        }
      }
    },
    "/user/change-password": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "修改密码",
        "description": "修改当前用户密码，修改后所有旧 Token 将失效",
        "operationId": "changePassword",
        "parameters": [
          {
            "name": "oldPassword",
            "in": "query",
            "description": "旧密码",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "newPassword",
            "in": "query",
            "description": "新密码",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refreshToken",
            "in": "query",
            "description": "Refresh Token",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/user/cancel": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "注销账号",
        "description": "用户自行注销账号（软删除），保留历史评论和评分，但无法再登录",
        "operationId": "cancelAccount",
        "parameters": [
          {
            "name": "refreshToken",
            "in": "query",
            "description": "Refresh Token",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/public/{userId}": {
      "get": {
        "tags": [
          "用户管理"
        ],
        "summary": "获取公开用户信息",
        "description": "根据用户ID获取公开的用户信息（不包含敏感信息如邮箱、角色等）",
        "operationId": "getPublicUserInfo",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "用户ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPublicUserVO"
                }
              }
            }
          }
        }
      }
    },
    "/user/info": {
      "get": {
        "tags": [
          "用户管理"
        ],
        "summary": "获取当前用户信息",
        "description": "根据请求头中的 Token 获取当前登录用户的详细信息",
        "operationId": "getInfo",
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultUserVO"
                }
              }
            }
          }
        }
      }
    },
...
    "/history/my": {
      "get": {
        "tags": [
          "浏览历史管理"
        ],
        "summary": "获取我的浏览历史",
        "description": "分页获取当前用户的浏览历史记录",
        "operationId": "getMyHistory",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页数量",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageInfoMyFavoriteVO"
                }
              }
            }
          }
        }
      }
    },
    "/history/count": {
      "get": {
        "tags": [
          "浏览历史管理"
        ],
        "summary": "获取浏览历史总数",
        "description": "统计当前用户的浏览历史总数",
        "operationId": "getHistoryCount",
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultInteger"
                }
              }
            }
          }
        }
      }
    },
...
    "/history/delete": {
      "delete": {
        "tags": [
          "浏览历史管理"
        ],
        "summary": "删除单条浏览历史",
        "description": "删除指定的浏览历史记录",
        "operationId": "deleteHistory",
        "parameters": [
          {
            "name": "historyId",
            "in": "query",
            "description": "历史记录ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/history/clear": {
      "delete": {
        "tags": [
          "浏览历史管理"
        ],
        "summary": "清空浏览历史",
        "description": "清空当前用户的所有浏览历史",
        "operationId": "clearHistory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/history/batch": {
      "delete": {
        "tags": [
          "浏览历史管理"
        ],
        "summary": "批量删除浏览历史",
        "description": "批量删除多条浏览历史记录",
        "operationId": "deleteBatchHistory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "historyIds": {
                    "type": "array",
                    "description": "历史记录ID列表",
                    "items": {
                      "type": "integer",
                      "format": "int64"
                    }
                  },
                  "user": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户ID (唯一账号)",
            "example": "movie_fan_01"
          },
          "nickname": {
            "type": "string",
            "description": "用户昵称",
            "example": "电影狂热者"
          },
          "avatar": {
            "type": "string",
            "description": "用户头像 URL",
            "example": "http://localhost:8080/images/default_avatar.png"
          },
          "url": {
            "type": "string",
            "description": "个人主页/豆瓣链接",
            "example": "https://www.douban.com/people/xxxx"
          },
          "email": {
            "type": "string",
            "description": "电子邮箱",
            "example": "user@example.com"
          },
          "role": {
            "type": "integer",
            "description": "用户角色 (0:管理员, 1:普通用户)",
            "format": "int32",
            "example": 1
          },
          "status": {
            "type": "integer",
            "description": "账号状态 (0:正常, 1:被禁用/冻结)",
            "format": "int32",
            "example": 0
          },
          "passwordVersion": {
            "type": "integer",
            "description": "密码版本号，用于失效旧 Token（修改密码后递增）",
            "format": "int32",
            "example": 1
          },
          "createTime": {
            "type": "string",
            "description": "账号创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "最后更新时间",
            "format": "date-time"
          }
        },
        "description": "用户实体"
      },
      "ResultString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200表示成功，其他为错误码)",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "Success"
          },
          "data": {
            "type": "string",
            "description": "业务数据载荷"
          }
        },
        "description": "统一 API 响应结构"
      },
...
      "RegisterDTO": {
        "required": [
          "id",
          "nickname",
          "password"
        ],
        "type": "object",
        "properties": {
          "id": {
            "maxLength": 20,
            "minLength": 4,
            "type": "string",
            "description": "用户账号 (唯一标识)",
            "example": "movie_fan_01"
          },
          "password": {
            "maxLength": 2147483647,
            "minLength": 6,
            "type": "string",
            "description": "密码",
            "example": "123456"
          },
          "nickname": {
            "type": "string",
            "description": "用户昵称",
            "example": "电影狂热者"
          },
          "email": {
            "type": "string",
            "description": "电子邮箱",
            "example": "test@movie.com"
          },
          "url": {
            "type": "string",
            "description": "个人主页/豆瓣主页链接",
            "example": "https://douban.com/people/xxx"
          }
        },
        "description": "用户注册请求参数"
      },
      "LoginDTO": {
        "required": [
          "id",
          "password"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户账号",
            "example": "movie_fan_01"
          },
          "password": {
            "type": "string",
            "description": "用户密码",
            "example": "123456"
          }
        },
        "description": "用户登录请求参数"
      },
      "ResultUserVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200表示成功，其他为错误码)",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/UserVO"
          }
        },
        "description": "统一 API 响应结构"
      },
      "UserVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户ID",
            "example": "movie_fan_01"
          },
          "nickname": {
            "type": "string",
            "description": "用户昵称",
            "example": "电影狂热者"
          },
          "avatar": {
            "type": "string",
            "description": "用户头像URL (自动拼接域名)",
            "example": "http://localhost:8080/images/default.png"
          },
          "url": {
            "type": "string",
            "description": "个人主页链接",
            "example": "https://douban.com/people/xxx"
          },
          "email": {
            "type": "string",
            "description": "电子邮箱",
            "example": "test@movie.com"
          },
          "role": {
            "type": "integer",
            "description": "用户角色 (0:管理员, 1:普通用户)",
            "format": "int32",
            "example": 1
          },
          "accessToken": {
            "type": "string",
            "description": "JWT 访问令牌 (短效)",
            "example": "eyJhbGciOiJIUzUxMiJ9..."
          },
          "refreshToken": {
            "type": "string",
            "description": "JWT 刷新令牌 (长效)",
            "example": "eyJhbGciOiJIUzUxMiJ9..."
          },
          "receivedLikes": {
            "type": "integer",
            "description": "用户获得的赞数",
            "format": "int32",
            "example": 1024
          },
          "commentCount": {
            "type": "integer",
            "description": "用户发布的影评数",
            "format": "int32",
            "example": 50
          }
        },
        "description": "登录成功后返回的用户信息"
      },
...
      "PublicUserVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户ID",
            "example": "movie_fan_01"
          },
          "nickname": {
            "type": "string",
            "description": "用户昵称",
            "example": "电影狂热者"
          },
          "avatar": {
            "type": "string",
            "description": "用户头像URL",
            "example": "http://localhost:8080/images/default.png"
          },
          "url": {
            "type": "string",
            "description": "用户豆瓣主页链接",
            "example": "https://douban.com/people/xxx"
          },
          "receivedLikes": {
            "type": "integer",
            "description": "用户获得的赞数",
            "format": "int32",
            "example": 1024
          },
          "commentCount": {
            "type": "integer",
            "description": "用户发布的影评数",
            "format": "int32",
            "example": 50
          }
        },
        "description": "公开用户信息（不包含敏感信息）"
      },
      "ResultPublicUserVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200表示成功，其他为错误码)",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/PublicUserVO"
          }
        },
        "description": "统一 API 响应结构"
      },
...
      "PageInfoMyFavoriteVO": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MyFavoriteVO"
            }
          },
          "pageNum": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "startRow": {
            "type": "integer",
            "format": "int64"
          },
          "endRow": {
            "type": "integer",
            "format": "int64"
          },
          "pages": {
            "type": "integer",
            "format": "int32"
          },
          "prePage": {
            "type": "integer",
            "format": "int32"
          },
          "nextPage": {
            "type": "integer",
            "format": "int32"
          },
          "isFirstPage": {
            "type": "boolean"
          },
          "isLastPage": {
            "type": "boolean"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          },
          "navigatePages": {
            "type": "integer",
            "format": "int32"
          },
          "navigatepageNums": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "navigateFirstPage": {
            "type": "integer",
            "format": "int32"
          },
          "navigateLastPage": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "业务数据载荷"
      },
      "ResultPageInfoMyFavoriteVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200表示成功，其他为错误码)",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/PageInfoMyFavoriteVO"
          }
        },
        "description": "统一 API 响应结构"
      },
      "ResultInteger": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200表示成功，其他为错误码)",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "Success"
          },
          "data": {
            "type": "integer",
            "description": "业务数据载荷",
            "format": "int32"
          }
        },
        "description": "统一 API 响应结构"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}